// src/components/v5/common/ActionSidebarRow/enums.ts
export enum ActionSidebarRowFieldNameEnum {
  ACTION_TYPE = 'actionType',
  DECISION_METHOD = 'decisionMethod',
  CREATED_IN = 'createdIn',
  DESCRIPTION = 'description',
  FROM = 'from',
  TO = 'to',
  RECIPIENT = 'recipient',
  AMOUNT = 'amount',
  ADDITIONAL_PAYMENTS = 'additionalPayments',
  CLAIM_DELAY = 'claimDelay',
  TOKEN = 'token',
  PERCENT = 'percent',
  MILESTONE = 'milestone',
  STARTS = 'starts',
  ENDS = 'ends',
  AMOUNT_PER = 'amountPer',
  LIMIT = 'limit',
  TOKENS_TO_CREATE = 'tokensToCreate',
  APPROVED_TOKENS = 'approvedTokens',
  TEAM = 'team',
  TEAM_NAME = 'teamName',
  TEAM_PURPOSE = 'teamPurpose',
  TEAM_COLOUR = 'teamColour',
  COLONY_NAME = 'colonyName',
  COLONY_LOGO = 'colonyLogo',
  COLONY_DESCRIPTION = 'colonyDescription',
  CURRENT_VERSION = 'currentVersion',
  NEW_VERSION = 'newVersion',
}